ci/requirements.txt
jsonschema
tqdm
munch
requests
pillow
